frontend/requirements.txt
# Core dependencies
kivy==2.2.1
kivymd==1.1.1
requests==2.31.0
websockets==12.0
pillow==10.1.0
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1
certifi==2023.11.17
urllib3==2.1.0
geopy==2.4.1

# iOS specific dependencies
pyobjus==1.2.1; platform_system=="Darwin"

# Android specific dependencies
android-permissions==1.1.1; platform_system=="Linux"
plyer==2.1.0

# Development tools
black==23.11.0
flake8==6.1.0
pytest==7.4.3
pytest-kivy==0.3.1
